Feat: Split av BeseEntitet til BaseCreateableEntitet som kan brukes med Immutable entiteter som ikke trenger endret_tid og endret_av
