+++ behandlingslager/domene/src/main/java/no/nav/foreldrepenger/behandlingslager/BaseEntitet.java
@@ -6,29 +6,16 @@ import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.MappedSuperclass;
-import javax.persistence.PrePersist;
 import javax.persistence.PreUpdate;
 
 import no.nav.foreldrepenger.behandlingslager.diff.DiffIgnore;
-import no.nav.vedtak.sikkerhet.context.SubjectHandler;
 
 /**
  * En basis {@link Entity} klasse som håndtere felles standarder for utformign
- * av tabeller (eks. sporing av hvem som har opprettet eller oppdatert en rad,
- * og når).
+ * av tabeller (eks. sporing av hvem som har oppdatert en rad og når).
  */
 @MappedSuperclass
-public abstract class BaseEntitet implements Serializable {
-
-    private static final String BRUKERNAVN_NÅR_SIKKERHETSKONTEKST_IKKE_FINNES = "VL";
-
-    @DiffIgnore
-    @Column(name = "opprettet_av", nullable = false, updatable = false)
-    private String opprettetAv;
-
-    @DiffIgnore
-    @Column(name = "opprettet_tid", nullable = false, updatable = false)
-    private LocalDateTime opprettetTidspunkt; // NOSONAR
+public abstract class BaseEntitet extends BaseCreateableEntitet implements Serializable {
 
     @DiffIgnore
     @Column(name = "endret_av")
@@ -38,32 +25,14 @@ public abstract class BaseEntitet implements Serializable {
     @Column(name = "endret_tid")
     private LocalDateTime endretTidspunkt; // NOSONAR
 
-    @PrePersist
-    protected void onCreate() {
-        this.opprettetAv = opprettetAv != null ? opprettetAv : finnBrukernavn();
-        this.opprettetTidspunkt = opprettetTidspunkt != null ? opprettetTidspunkt : LocalDateTime.now();
-    }
-
     @PreUpdate
     protected void onUpdate() {
         endretAv = finnBrukernavn();
         endretTidspunkt = LocalDateTime.now();
     }
 
-    public String getOpprettetAv() {
-        return opprettetAv;
-    }
-
-    public void setOpprettetAv(String opprettetAv) {
-        this.opprettetAv = opprettetAv;
-    }
-
     public void setEndretAv(String endretAv) {
         this.endretAv = endretAv;
-    }
-
-    public LocalDateTime getOpprettetTidspunkt() {
-        return opprettetTidspunkt;
     }
 
     public String getEndretAv() {
@@ -78,17 +47,4 @@ public abstract class BaseEntitet implements Serializable {
         this.endretTidspunkt = endretTidspunkt;
     }
 
-    private static String finnBrukernavn() {
-        String brukerident = SubjectHandler.getSubjectHandler().getUid();
-        return brukerident != null ? brukerident : BRUKERNAVN_NÅR_SIKKERHETSKONTEKST_IKKE_FINNES;
-    }
-
-    /**
-     * Kan brukes til å eksplisitt sette opprettet tidspunkt, f.eks. ved migrering
-     * av data fra et annet system. Ivaretar da opprinnelig tidspunkt istdf å sette
-     * likt now().
-     */
-    public void setOpprettetTidspunkt(LocalDateTime opprettetTidspunkt) {
-        this.opprettetTidspunkt = opprettetTidspunkt;
-    }
 }
